Add wrong solution to Exercise 3 wordFitsTemplate
It occurred to me that if I pulled the letters out of the template and
added them to the hand, I could just use formableBy with that expanded
hand to see if the given string was possible. I failed to notice that
this doesn't account for matches which are too short to complete the
pattern, as in the third test case, which I'm glad was included, or I
would have believed my incorrect solution for longer than I did.