Correct solution for Exercise 3 wordFitsTemplate
This is a lot more exhaustive than my previous 2 attempts, and I think
correct. An empty template with an empty string is a match. If each has
at least one character, then either they match, and we recurse on the
tails of each, or the template's head is a '?' and the string's head is
a character in our hand, in which case we remove it from the hand and
recurse with tails of the template and string. If we run out of both the
template and string simultaneously, the first pattern matches, yielding
True, and failing all of the above, we get a False.